M1: atk/frame -pi; retype Enumerate rock long->void*, drop laundering casts tree-wide
FossilOrigin-Name: 4dc01c4f70f1d6271e6d33887c062e885b52d6247d4bce1bdba3f14ddbc32f16

+++ src/atk/extensions/compile.c
@@ -954,7 +954,7 @@ static struct frame *FindByView(view)
     struct view *view;
 {
 
-    return frame_Enumerate(ViewEqual, (long) view);
+    return frame_Enumerate(ViewEqual, view);
 }
 
 /* Find a window other that the one that contains this inset.  Create one if we have to. */
@@ -972,7 +972,7 @@ static struct view *PutInAnotherWindow(view, buffer, forceWindow)
     myInfo.otherFrame = NULL;
     myInfo.bestFrame = NULL;
     myInfo.myBuffer = buffer;
-    frame_Enumerate(FrameFinder, (long) &myInfo);
+    frame_Enumerate(FrameFinder, &myInfo);
     frame = (myInfo.bestFrame != NULL) ? myInfo.bestFrame : ((myInfo.otherFrame != NULL) ? myInfo.otherFrame : NULL);
     if (frame == NULL) {
